+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1275](https://github.com/alleslabs/celatone-frontend/pull/1275) Add Initia App Menu from submodules
+
 ### Bug fixes
 
 ## v1.10.1